+++ CHANGELOG.md
@@ -32,3 +32,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Fix profile card layout issues on mobile.
+
+# [1.0.4] - 2020-12-15
+
+### Added
+
+- Add cursus name to profile card.
+- Add `Piscine C décloisonnée` to cursuses list.
+
+## Changed
+
+- Add `Pisciner` tag to piscine candidates.